Factor-out pieces of lax_numpy.py
